Extend parameters in preparation of '_get' and '_set' patches
Extends the parameters by filling in missing data that is known
beforehand (see the rotation-related hunk) and adding a new
'prop_type' field to allow for better type handling. While this
info won't really be used until '_get' and '_set' appear, the
'get_param' function has been updated to take advantage of them
in the meantime.