Add ability to match input aspect ratio to the current output ratio
Adds an optional "KeepShape" paramater to the MapToOutput command.
If provided, the Area property is updated to match the aspect ratio
of the new output. If not provided, the Area property is reset to
the full tablet dimensions.

This commit causes the MapToOutput command to break existing scripts
that rely on the Area property for device calibration. It is not
expected that this will result in actual problems since calibration
utilities do not currently handle multi-head setups properly anyway.