Treat KEY_PROG1, KEY_PROG2, and KEY_PROG3 as buttons
The driver does not currently pass through KEY_* events as keyboard
presses, and the modifications to make it do so are outside the
scope of this patch set. Until the driver is patched to change this
oddity, we treat the three KEY_* the 24HD sends from the kernel the
same as BTN_* events.