De-duplicate code for wheel/strip scrolling
Replaces sendWheelStripEvents and getWheelButton with several
small functions to reduce unnecessary code duplication.

As a side-effect, it is now possible for the driver to handle
simultaneous scrolls on multiple axes. Previously, getWheelButton
would go through all the axes but was limited to returning a
single button/action.

Changes from v3:
 * Add tests for getScrollDelta and getWheelButton
 * Fix bugs in getScrollDelta doc (wrap is max, not max+1)
 * Fix bugs in getScrollDelta implementation (log2(0) is undef)