Partial unification of 'scroll_keys' and 'keys' into 'actions'
Again, same type of data, but stored and handled seperately.
This code (mostly) unifies the two into a linked list of WacomAction
elements that define the up and down actions to take.

This does not touch wcmXCommand.c, where external code can set the
actions. That's going to take some thinking to get right.

I'll probably want to re-think the 'up' and 'down' variable names
in WacomAction. A button press is thought of as a "down" event,
even though when it occurs, the delta is "up". The semantic clash
is sure to cause a little confusion as things stand...