Find appropriate value for 'nbuttons' instead of assuming
Some Bamboo tablets have buttons mapped to "Forward" and "Back"
(X11 buttons 8 and 9) by default. Combined with having less than
4 physical buttons, this results in us not allocating enough
space in the Actions property to change their associated actions.

This patch scans the physical-to-X11 button mapping to find the
largest used X11 button number, and use *that* as the value for
'nbuttons'.