BRANCH DESCRIPTION - Unify all *_keys (action) data in one place
Right now the driver has several places to store data about the
action associated with a specific event. If an event comes from
a wheel (absolute or relative), we consult 'wheel_keys'. If instead
it comes from a touch strip, we consult 'strip_keys'. Finally, if
the event comes from a button, we consult 'keys'.

All three locations store the same data -- the action to perform
given an event. This branch aims to create a single location for
storage of action data, with the ability to find the appropriate
action given the event details.