Unify wheel_keys and strip_keys into 'scroll_keys'
Both wheel_keys and strip_keys store the exact same information:
actions to perform when its determined that the underlying axis
is scrolling. This patch unifies the two so that all scroll actions
are contained in a single variable.

To maintain the same interface to existing X programs, some pointer
magic is required in wcmXCommand.c.