Define pad and cursor default buttons based on kernel data
Instead of setting up a static mapping which "directly" associates
flags in the button int with X11 buttons, we determine the proper
X11 button to use based on the kernel event recieved. Normally
these are the same, but in the case of "mouse" buttons (which have
are now also stored in usbdata->padkey_code) don't work properly
since there are often "holes" in the button association.