Change gestures code to be Android-like
Tap: Reposition cursor
Move: Move cursor
Hold: Mouse down
Hold+Move: Drag cursor

This is accomplished by allowing LCDs to enter PREDRAG mode, which
is a staging area to determine if this is a drag or not. If still
in PREDRAG mode when the timer fires, the mouse button is pressed
and DRAG mode is entered. The PREDRAG mode can be exited prematurely
by moving the finger a minimum distance from its starting point or
by starting a multitouch gesture.

To pull this off, we need to only enter PREDRAG mode on the inital
touch, which means modifying wcmSingleFingerPress to have access
to historical states. Also, to ensure the pointer is repositioned
immediately (so taps work as listed), we need to have the driver
send events on PREDRAG as well as DRAG (and the implicit NONE).