WIP: Use oldState.proximity instead of oldCursorHwProx
oldCursorHwProx is used to determine if the cursor has just
entered proximity, so that the value of wcmMaxCursorDist
can be appropriately set (since Protocol 4 and 5 differ in
their definition of "distance"). There is no need for a
dedicated variable, however, since oldState.proximity is
available.

NOTE - oldState is not tied to a specific channel, so
this isn't strictly the same. If a second tool (esp. on
a dual-track Intuos, but maybe even a touch device) goes
out of prox, we could potentially have oldState.proximity=0
and yet have the cursor (on a different channel) still in
prox.

What we *REALLY* should do is set wcmMaxCursorDist beforehand :p
IIRC, this is a protocol distinction, and isn't something that
needs to be updated as tools come and go. Distance for a pen
and a puck are both going to move in the same "direction"
provided the protocol is the same, right? Then we don't even
need oldCursorHwProx (or an analogue) at all!