WIP: Allow EMR to grab pointer control only when pointer is idle
This patch changes the how pointer access is arbitrated. Before,
any EMR tool would be immediatley granted control over the
pointer, preempting any tool that may have had it beforehand.
This could lead to "fighting" if two EMR tools were used
simultaneously.

Now, the driver allows a tool to maintain control of the pointer
so long as it is actively generating events. If, for instance,
a pen is brought in proximity of one tablet, a second pen
brought into prox of a second tablet will no longer cause
fighting over the pointer. Deciding if a tool is active based on
its last event timestamp rather than proximity allows a pen
to be used even if a puck has been left on another tablet.

*** NOTE ***

There's been back-and-forth on the issue of this kind of cross-
device EMR arbitration on Windows and Mac. For now, we're going
to continue to use existing behavior since nobody has complained.