WIP: Allow an event with a serial number to take over an existing serial-less slot
Tablets either report serial numbers or they do not. AES tablets report serial
numbers, but there is a window of time after entering proximity when we get
several reports with no serial number. Initial pen state will be stored with
a serial number of "1" (just like for tablets that report no serial number),
but later ones will be sent to a different channel with the correct serial
number. If significant pen state information (e.g. the fact that a side-switch
is pressed) is stored in the channel for serial "1", then this information
will effectively disappear once updates start going to the channel for the
real serial.

By allowing events that have serial numbers to take over existing serial-
less channels, we can ensure that the full history is preserved. Since
there's no such thing as a dual-track tablet which reports some pens
without a serial and some pens with a serial, this should be safe.

https://sourceforge.net/p/linuxwacom/bugs/339/