Use git-version-gen to set AC_INIT version data
The AC_INIT macro is used to define basic package information such as the
version number. At the moment, we have to manually bump this number as
part of the release process. The version number information is already
available from git, however, meaning that we can instead use a script like
git-version-gen to fill it in.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>