Trigger scroll and zoom gestures immediately after they are detected
The amount you have to move your fingers to start a gesture is quite a
larger than the amount you have to move them to continue to gesture.
This makes the smooth use of gestures difficult since it can be a little
unpredictable exactly when they finally kick in. By triggering the zoom
and scroll gestures at the same distance as it takes to continue them,
the gestures feel much more smooth to execute.

This commit removes the various delays that prevent gestures from being
executed the moment the configured scroll or zoom distance is achieved.